Remove circular Spring bean dependencies
